+++ src/assets/Computer/Desktop/Icon/notepad.png
