+++ src/components/Computer/Desktop/Window/MainArea/mainArea.js
