+++ src/components/Computer/Taskbar/TaskbarItem/taskbar-item.js
@@ -1,0 +1,17 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+
+import './taskbar-item.css';
+
+const taskbarItem = props =>(
+    <div className="taskbar-item" onClick={() => props.onClick()}>
+        {this.props.name}
+    </div>
+)
+
+taskbarItem.propTypes = {
+    onClick: PropTypes.func.isRequired,
+    name: PropTypes.string.isRequired,
+}
+
+export default taskbarItem;
