--- src/components/Desktop/desktop.css
@@ -1,4 +1,0 @@
-.desktop {
-    width: 100%;
-    height: 100%;
-}
