--- src/components/Desktop/workDesk.css
@@ -1,7 +1,0 @@
-.work-desk {
-    top:0;
-    left: 0;
-    right: 0;
-    bottom: 50px;
-    position: fixed;
-}
