+++ src/containers/Computer/Desktop/Window/programWindow.js
@@ -1,11 +1,11 @@
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
-import { closeProgram, hideProgram, fullscreenProgram } from '../../actions/openProgram';
-import WindowBar from './WindowBar/windowBar';
-import MenuBar from './MenuBar/menuBar';
-import MainArea from './MainArea/mainArea';
-import Draggable from 'react-draggable';
 import PropTypes from 'prop-types';
+import Draggable from 'react-draggable';
+import { closeProgram, hideProgram, fullscreenProgram } from '../../../../actions/openProgram';
+import WindowBar from '../../../../components/Computer/Desktop/Window/WindowBar/windowBar';
+import MenuBar from '../../../../components/Computer/Desktop/Window/MenuBar/menuBar';
+import MainArea from '../../../../components/Computer/Desktop/Window/MainArea/mainArea';
 import './programWindow.css';
 
 class ProgramWindow extends Component {
