+++ src/store.js
@@ -1,8 +1,8 @@
 import {createStore, compose, applyMiddleware} from 'redux';
 import {createLogger} from 'redux-logger';
 import rootReducer from './reducers/index';
-import MyComputerLogo from './components/Desktop/my_computer.png';
-import NotepadLogo from './components/Desktop/Notepad.png';
+import MyComputerLogo from './assets/Computer/Desktop/Icon/my_computer.png';
+import NotepadLogo from './assets/Computer/Desktop/Icon/Notepad.png';
 
 const loggerMiddleware = createLogger();
